requirements.txt
Django==3.2.18
django-ckeditor==6.3.2
